requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aiosqlite==0.15.0
async-timeout==3.0.1
asyncpg==0.21.0
atlastk==0.11.1
attrs==20.1.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
iso8601==0.1.12
loguru==0.5.1
multidict==4.7.6
PyPika==0.38.0
python-dotenv==0.14.0
pytz==2020.1
tortoise-orm==0.16.14
typing-extensions==3.7.4.2
yarl==1.5.1
